core: Rename rt::io to rt::rtio
This is an internal interface. I want to use rt::io for public interfaces.
