+++ src/libcore/rt/rtio.rs
